requirements.test.txt
coverage==7.4.3
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-homeassistant-custom-component==0.13.109
requests
requests_oauthlib
